+++ Assets/Settings/ScriptableObject/CodeEditorSettings.asset
@@ -10,7 +10,7 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 2628e7e472a5d9744bb4ebbe7085f74d, type: 3}
-  m_Name: CpdeEditorSettings
+  m_Name: CodeEditorSettings
   m_EditorClassIdentifier: 
   FontSize: 0.5
   LineHeight: 0.72
