+++ Assets/Sources/Entities/Shared.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6d5ad6841cfe573439d91c64be785785
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
