--- Assets/Sources/InterfaceAdapters/CodeEditor/ITextAreaLayoutPresenter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a52c4febf9feba849a3b904a8a36c3fa
