--- Assets/Sources/InterfaceAdapters/CodeEditor/TextAreaLayoutPresenter.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 2c18434deb3e449459354beaf4b342a1
