+++ Assets/Sources/InterfaceAdapters/Controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 7302e98aaf415a847a5194fcd8e0d562
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
