+++ Assets/Sources/InterfaceAdapters/Controller/IGetInputFieldText.cs
@@ -1,0 +1,10 @@
+using UnityLike.Entities.CodeEditor;
+
+namespace UnityLike.InterfaceAdapters.Controller
+{
+    public interface IGetInputFieldText
+    {
+        string GetInputFieldText(CodeEditorBlock block);
+    }
+
+}
