+++ Assets/Sources/InterfaceAdapters/Controller/IGetInputFieldText.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: bcd34a78f377df24490a2eed0740d4d2
