+++ Assets/Sources/InterfaceAdapters/Controller/ITextAreaInput.cs
@@ -1,0 +1,11 @@
+using System;
+
+using UnityLike.Entities.CodeEditor;
+
+namespace UnityLike.InterfaceAdapters.Controller
+{
+    public interface ITextAreaInput
+    {
+        event Action<CodeEditorBlock, string> OnTextAreaInputChanged;
+    }
+}
