+++ Assets/Sources/InterfaceAdapters/Controller/ITextAreaInput.cs.meta
