+++ Assets/Sources/InterfaceAdapters/Presenter.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 457cb1ca050611046ad70f0e4cb368a4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
