+++ Assets/Sources/InterfaceAdapters/Presenter/ITextAreaLayoutPresenter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5b1487f74b03e7745ad01faaafd169ef
