+++ Assets/Sources/InterfaceAdapters/Presenter/ITextAreaView.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5d049766ab6173142afc394438054e8b
