+++ Assets/Sources/InterfaceAdapters/Presenter/TextAreaLayoutPresenter.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 5f2e3604cbbc158429a2f6d7d0958bc5
